# [ML] inference performance optimizations