--- a/x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/ml/utils/MapHelper.java
+++ b/x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/ml/utils/MapHelper.java
@@ -6,8 +6,10 @@
 package org.elasticsearch.xpack.core.ml.utils;
 
 import org.elasticsearch.common.Nullable;
+import org.elasticsearch.common.Strings;
 
-import java.util.Arrays;
+import java.util.Collection;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Stack;
 
@@ -57,23 +59,49 @@ public final class MapHelper {
      *
      * Instead we lazily create potential paths once we know that they are possibilities.
      *
-     * @param path Dot delimited path containing the field desired
+     * @param path Dot delimited path containing the field desired. Assumes that the path contains no empty strings
      * @param map The {@link Map} map to dig
      * @return The found object. Returns {@code null} if not found
      */
     @Nullable
     public static Object dig(String path, Map<String, Object> map) {
         // short cut before search
-        if (map.keySet().contains(path)) {
-            return map.get(path);
+        Object obj = map.get(path);
+        if (obj != null) {
+            return obj;
         }
-        String[] fields = path.split("\\.");
-        if (Arrays.stream(fields).anyMatch(String::isEmpty)) {
-            throw new IllegalArgumentException("Empty path detected. Invalid field name");
-        }
+        String[] fields = Strings.delimitedListToStringArray(path, ".");
         Stack<PotentialPath> pathStack = new Stack<>();
         pathStack.push(new PotentialPath(map, 0));
         return explore(fields, pathStack);
+    }
+
+    /**
+     * Collapses dot delimited fields so that the map is a single layer.
+     *
+     * Example:
+     * {
+     *     "a" :{"b": {"c": {"d" : 2}}}
+     * }
+     * becomes:
+     * {
+     *     "a.b.c.d": 2
+     * }
+     *
+     * @param map The map that has nested and/or collapsed paths
+     * @param pathsToCollapse The desired paths to collapse
+     * @return A fully collapsed map
+     */
+    public static Map<String, Object> dotCollapse(Map<String, Object> map, Collection<String> pathsToCollapse) {
+        // default load factor is 0.75 (3/4).
+        Map<String, Object> collapsed = new HashMap<>(((pathsToCollapse.size() * 4)/3) + 1);
+        for (String path : pathsToCollapse) {
+            Object dug = dig(path, map);
+            if (dug != null) {
+                collapsed.put(path, dug);
+            }
+        }
+        return collapsed;
     }
 
     @SuppressWarnings("unchecked")
@@ -95,8 +123,11 @@ public final class MapHelper {
                 }
                 endPos++;
             }
-            if (candidateKey != null && map.containsKey(candidateKey)) { //exit early
-                return map.get(candidateKey);
+            if (candidateKey != null) { // exit early
+                Object val = map.get(candidateKey);
+                if (val != null) {
+                    return val;
+                }
             }
         }