--- a/x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/ml/inference/trainedmodel/tree/TreeNodeTests.java
+++ b/x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/ml/inference/trainedmodel/tree/TreeNodeTests.java
@@ -12,9 +12,7 @@ import org.elasticsearch.xpack.core.ml.job.config.Operator;
 import org.junit.Before;
 
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.List;
 
 import static org.hamcrest.Matchers.equalTo;
 
@@ -85,9 +83,9 @@ public class TreeNodeTests extends AbstractSerializingTestCase<TreeNode> {
 
     public void testCompare() {
         expectThrows(IllegalArgumentException.class,
-            () -> createRandomLeafNode(randomDouble()).compare(Collections.singletonList(randomDouble())));
+            () -> createRandomLeafNode(randomDouble()).compare(new double[]{randomDouble()}));
 
-        List<Double> featureValues = Arrays.asList(0.1, null);
+        double[] featureValues = new double[]{0.1, Double.NaN};
         assertThat(createRandom(0, 2, 3, 0.0, 0, null).build().compare(featureValues),
             equalTo(3));
         assertThat(createRandom(0, 2, 3, 0.0, 0, Operator.GT).build().compare(featureValues),