This optimizes performance for inference models in the following ways:

 - When a model is loaded by the model loading service, feature extraction now only takes place at the top level model (most cases, the ensemble)
 - Flattens out the field map before feature extraction. This allows us to only "dig" the desired fields only once.
 - changes some data structures to be actual arrays instead of unmodifiable lists

All this significantly reduces runtime along with some minor memory improvements.